requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-jose==3.3.0
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic-settings==2.5.2
openpyxl==3.1.5
email-validator==2.2.0
